Organize files, create simple tilling files, try new program.exe
