--- Result_Summary.md
@@ -1,20 +1,0 @@
-Using multi-threading for matrix filling in:
-Matrix size:                        1000 x 1000
-Dot Product:                        613595280
-Time elapsed to fill matrices:      1.52312
-Time elapsed to calculate product:  0.003542
--------------------------------------------------------------------------
-Using multi-threading for matrix filling in after introducing semaphore:
-Matrix size:                        1000 x 1000
-Dot Product:                        611764021
-Time elapsed to fill matrices:      1.52063
-Time elapsed to calculate product:  0.310049
--------------------------------------------------------------------------
-Using multi-threading for both matrix modifications:
-Matrix size:                        1000 x 1000
-Dot Product:                        617440260
-Time elapsed to fill matrices:      1.54707
-Time elapsed to calculate product:  1.60764
-
-DELETE EVERYTHING ABOVE
--------------------------------------------------------------------------
