+++ Results.md
@@ -1,0 +1,12 @@
+## Simple Mode
+Matrix size:                        1000 x 1000
+Result[0][0]:                       2312883
+Time elapsed to fill matrices:      22.1419
+Time elapsed to calculate product:  2.54393
+
+## Multithreading Mode
+Matrix size:                        1000 x 1000
+Result[0][0]:                       4663864
+Time elapsed to fill matrices:      5.58914
+Time elapsed to calculate product:  0.660256
+
