+++ include/MTMode.hpp
@@ -1,7 +1,7 @@
 #ifndef MTMODE_HPP
 #define MTMODE_HPP
 
-int calculateProduct(int threadID);
+int calculateProductMT(int threadID);
 void MTMode();
 
 #endif
