+++ include/SimpleTilling.hpp
@@ -1,0 +1,7 @@
+#ifndef SIMPLETILLING_HPP
+#define SIMPLETILLING_HPP
+
+int calculateProductSMPLTilling();
+void SimpleTilling();
+
+#endif
