+++ src/MTMode.cpp
@@ -6,44 +6,15 @@
 #include <filesystem>
 #include <chrono>
 #include <windows.h>
+#include "../include/MTMode.hpp"
+#include "../include/main.hpp"
 
 using namespace std;
 
-// Stores the one dimensional size of the matrix
-const int MATRIXLENGTH = 1000;
-// Creates matrices of MATRIXLENGTH in length and width!
-// Thus, the dimensions of the matrix is MATRIXLENGTH x MATRIXLENGTH
-int matrix1[MATRIXLENGTH][MATRIXLENGTH];
-int matrix2[MATRIXLENGTH][MATRIXLENGTH];
-int result[MATRIXLENGTH][MATRIXLENGTH];
-
-// Semaphore variable!
-HANDLE productLock;
-
-// Create a random number generator with an upper and lower bound! (inclusive)
-int randomNumberGenerator(int lower_bound, int upper_bound) {
-    random_device rd;
-    mt19937 gen(rd());
-
-    uniform_int_distribution<> distr(lower_bound, upper_bound);
-
-    return distr(gen);
-}
-
-void fillMatrices() {
-    // Fill up the matrices!
-    for (int i = 0; i < MATRIXLENGTH; i++) {
-        for (int j = 0; j < MATRIXLENGTH; j++) {
-            matrix1[i][j] = randomNumberGenerator(0, 99);
-            matrix2[i][j] = randomNumberGenerator(0, 99);
-        }
-    }
-}
-
 // Calculate multiplicative product of the two matrices: matrix1 and matrix2
-int calculateProduct() {
+int calculateProductMT(int threadID) {
     // Loop through rows
-    for(int i = 0; i < MATRIXLENGTH; i++) {
+    for(int i = threadID; i < MATRIXLENGTH; i += THREADCOUNT) {
         // Loop through columns
         for(int j = 0; j < MATRIXLENGTH; j++) {
             long temp = 0;
@@ -59,29 +30,35 @@ int calculateProduct() {
     return 0;
 }
 
-int main (int argc, char* argv[]) {
-    //Creates output file
-    ofstream outFile("Results.txt");
-
-    productLock = CreateSemaphore(nullptr, 1, 1, nullptr);
-
-    if(productLock == nullptr) {
-        cerr << "CreateSemaphore error!" << GetLastError() << endl;
-        return 1;
-    }
+// Acts as main for the multithreading mode!
+void MTMode() {
+    thread threads[THREADCOUNT];
 
     auto startFill = chrono::high_resolution_clock::now();
-    fillMatrices();
+    for(int i = 0; i < THREADCOUNT; i ++) {
+        threads[i] = thread(fillMatrices, i);
+    }
+
+    for(auto& th: threads) {
+        th.join();
+    }
     auto endFill = chrono::high_resolution_clock::now();
 
     chrono::duration<double> durationFill = endFill - startFill;
 
     auto startCalc = chrono::high_resolution_clock::now();
-    calculateProduct();
+    for(int i = 0; i < THREADCOUNT; i ++) {
+        threads[i] = thread(calculateProductMT, i);
+    }
+
+    for(auto& th: threads) {
+        th.join();
+    }
     auto endCalc = chrono::high_resolution_clock::now();
 
     chrono::duration<double> durationCalc = endCalc - startCalc;
 
+    outFile << "## Multithreading Mode" << endl;
     outFile << "Matrix size:                        " << 
     MATRIXLENGTH << " x " << MATRIXLENGTH << endl;
     outFile << "Result[0][0]:                       " << 
@@ -89,5 +66,5 @@ int main (int argc, char* argv[]) {
     outFile << "Time elapsed to fill matrices:      " << 
     durationFill.count() << endl;
     outFile << "Time elapsed to calculate product:  " << 
-    durationCalc.count() << endl;
+    durationCalc.count() << "\n" << endl;
 }
